--- src/components/NavBar/Navbar.tsx
@@ -1,116 +1,0 @@
-import { Trans } from '@lingui/macro'
-import { ChainSelector } from 'components/NavBar/ChainSelector'
-import { MenuDropdown } from 'components/NavBar/MenuDropdown'
-import * as styles from 'components/NavBar/NavBar.css'
-import { SearchBar } from 'components/NavBar/SearchBar'
-import { ShoppingBag } from 'components/NavBar/ShoppingBag'
-import Web3Status from 'components/Web3Status'
-import { NftVariant, useNftFlag } from 'featureFlags/flags/nft'
-import { Box } from 'nft/components/Box'
-import { Row } from 'nft/components/Flex'
-import { UniIcon } from 'nft/components/icons'
-import { ReactNode } from 'react'
-import { NavLink, NavLinkProps, useLocation } from 'react-router-dom'
-
-interface MenuItemProps {
-  href: string
-  id?: NavLinkProps['id']
-  isActive?: boolean
-  children: ReactNode
-}
-
-const MenuItem = ({ href, id, isActive, children }: MenuItemProps) => {
-  return (
-    <NavLink
-      to={href}
-      className={isActive ? styles.activeMenuItem : styles.menuItem}
-      id={id}
-      style={{ textDecoration: 'none' }}
-    >
-      {children}
-    </NavLink>
-  )
-}
-
-const PageTabs = () => {
-  const { pathname } = useLocation()
-  const nftFlag = useNftFlag()
-
-  const isPoolActive =
-    pathname.startsWith('/pool') ||
-    pathname.startsWith('/add') ||
-    pathname.startsWith('/remove') ||
-    pathname.startsWith('/increase') ||
-    pathname.startsWith('/find')
-
-  return (
-    <>
-      <MenuItem href="/swap" isActive={pathname.startsWith('/swap')}>
-        <Trans>Swap</Trans>
-      </MenuItem>
-      <MenuItem href="/tokens" isActive={pathname.startsWith('/tokens')}>
-        <Trans>Tokens</Trans>
-      </MenuItem>
-      {nftFlag === NftVariant.Enabled && (
-        <MenuItem href="/nfts" isActive={pathname.startsWith('/nfts')}>
-          <Trans>NFTs</Trans>
-        </MenuItem>
-      )}
-      <MenuItem href="/pool" id={'pool-nav-link'} isActive={isPoolActive}>
-        <Trans>Pool</Trans>
-      </MenuItem>
-    </>
-  )
-}
-
-const Navbar = () => {
-  const { pathname } = useLocation()
-  const isNftPage = pathname.startsWith('/nfts')
-
-  return (
-    <>
-      <nav className={styles.nav}>
-        <Box display="flex" height="full" flexWrap="nowrap" alignItems="stretch">
-          <Box className={styles.leftSideContainer}>
-            <Box as="a" href="#/swap" className={styles.logoContainer}>
-              <UniIcon width="48" height="48" className={styles.logo} />
-            </Box>
-            <Box display={{ sm: 'flex', lg: 'none' }}>
-              <ChainSelector leftAlign={true} />
-            </Box>
-            <Row gap="8" display={{ sm: 'none', lg: 'flex' }}>
-              <PageTabs />
-            </Row>
-          </Box>
-          <Box className={styles.middleContainer}>
-            <SearchBar />
-          </Box>
-          <Box className={styles.rightSideContainer}>
-            <Row gap="12">
-              <Box display={{ sm: 'flex', xl: 'none' }}>
-                <SearchBar />
-              </Box>
-              <Box display={{ sm: 'none', lg: 'flex' }}>
-                <MenuDropdown />
-              </Box>
-              {isNftPage && <ShoppingBag />}
-              <Box display={{ sm: 'none', lg: 'flex' }}>
-                <ChainSelector />
-              </Box>
-
-              <Web3Status />
-            </Row>
-          </Box>
-        </Box>
-      </nav>
-      <Box className={styles.mobileBottomBar}>
-        <PageTabs />
-        <Box marginY="4">
-          <MenuDropdown />
-        </Box>
-      </Box>
-    </>
-  )
-}
-
-export default Navbar
